Added method to get all sensor data from coodinates.
This is a combination of two previous methods: one to get sensors in a
specific radius, and another to get data from a specific sensor. This
method removes the requirement to handle multiple functions to construct
a single response and allows the function to pipeline the request to get
multiple sensors' data.

All utilitiy functions specific to the database package were moved to a
new file in that package called utils.go.

Improved the Redis slice to float slice conversion function so that it
can properly handle status messages created by pipelines.

Added function to extract IDs from Redis response structs.