Moved task timeout logic into task package.
Adds WrapTimeout function that takes a task and wraps the Run method
with a timeout.

Adds GetTTL and SkipStartup methods to Task interface. GetTTL returns
a duration for the maximum amount of time a task should run. SkipStartup
skips the task during the startup run.

Updated unit tests to test these new features.