Adding AQI forecast stream handlers.
These methods allow for the server to access data streams containing
computed data for predicting future AQI forecasts. This is a critical
system required to notify user to changing air quality conditions.

Added unit tests for utilities in redis package. These functions to the
majority of the data serialization from the Redis datastore and
realistically are the only part of the pipeline that needs to be tested.