Updated AQIStreamData.
Renamed AQIStreamData to NotificationStream. Struct now includes a UID
field that refers to a user in the sql database and MessageID which is a
reference to the message from which the Notification was serialized.
This allows the message to be later acknowleged.

Removed GetForecastsStreamForIDs as this method was not useful.

Added NotificationConsumerRead that allows clients to recieve
notifications from the datastore stream. Added method to acknowledge
notifications in datastore.

Modified utils.go to handle the above changes. Added check for Nil data
result (i.e. no outstanding notifications were found).

Updated utils unit tests to accomodate these new changes.