Added last_crossover to users table.
This field will be used to track when the AQI for a specific
user crossed over their threshold. This will enable accurate delivery of
notification.

Added convenience function to update this field in SQL package.

Added LastCrossover field to UserRequest struct and marked it as
non-serializable for json encoder/decoder.

Regenerated sql models package with new column.