Added web SSL.
This program now has the ability to automatically download and provision
SSL certificates from Let's Encrypt using the autocert package.

Added ssl options to viper as a subsection under web. This allows a user
to enable SSL and configure domains for SSL. It also allows the user to
specify an email for Let's Encrypt notifications.